cleanup(bench): single shared bench root, drop experiment-era wording
The keyed and recycle perf benches each defined an identical Krausest-shaped
root component + template (differing only in the each key) — now one
defineBenchRoot(eachKey) factory in __test-utils__/list-harness.ts, so the
two variants cannot drift apart. Harness headers rewritten for their merged
role (the experiment-branch coordination notes were only meaningful while
six prototype branches shared the file).

Verified select against the production deploy preview (select/deselect/
switch rows all correct, the only console error is the pre-existing
rehydration fallback also present on master's production site); suite green,
gated benches reproduce prior numbers.
